Move process spawning to Core method
That makes it impossible to spawn a process before creating a Core instance.
Otherwise one could start running a process and because unix signal handler
is not yet installed we could miss that the child process terminated.

Also make method general over owned or mut borrowed ref to Command, for
user convenience.